refactor: split components; add loading
