+++ src/routes/_social-workers/editais/-data.tsx
@@ -1,0 +1,105 @@
+import type { Student } from "@/components/students/student-table";
+
+
+export const students: Student[] = [
+  {
+    id: 1,
+    cpf: "123.456.789-01",
+    nome: "Ana Souza",
+    matricula: "20250101",
+    status: "Pendente",
+    progresso: 0,
+    documentos: 11,
+    dataInscricao: "2025-02-06T13:19:00",
+  },
+  {
+    id: 1,
+    cpf: "123.456.789-07",
+    nome: "Lucas Martins",
+    matricula: "20250106",
+    status: "Pendente",
+    progresso: 0,
+    documentos: 22,
+    dataInscricao: "2025-02-05T16:00:00",
+  },
+  {
+    id: 1,
+    cpf: "772.910.068-13",
+    nome: "Marli Clarice Silveira",
+    matricula: "20240102",
+    status: "Pendente",
+    progresso: 0,
+    documentos: 22,
+    dataInscricao: "2025-02-05T14:41:00",
+  },
+  {
+    id: 1,
+    cpf: "573.850.954-45",
+    nome: "Severino Yuri Araújo",
+    matricula: "22111481",
+    status: "Pendente",
+    progresso: 0,
+    documentos: 22,
+    dataInscricao: "2025-02-05T14:21:00",
+  },
+  {
+    id: 1,
+    cpf: "123.456.789-05",
+    nome: "Fernanda Costa",
+    matricula: "20250105",
+    status: "Em Análise",
+    progresso: 20,
+    documentos: 10,
+    dataInscricao: "2025-02-04T09:45:00",
+  },
+  {
+    id: 1,
+    cpf: "123.456.789-03",
+    nome: "Beatriz Lima",
+    matricula: "20250103",
+    status: "Recurso",
+    progresso: 40,
+    documentos: 16,
+    dataInscricao: "2025-02-04T13:38:00",
+  },
+  {
+    id: 1,
+    cpf: "123.456.789-02",
+    nome: "João Pereira",
+    matricula: "20250102",
+    status: "Deferido",
+    progresso: 100,
+    documentos: 28,
+    dataInscricao: "2025-02-03T13:21:00",
+  },
+  {
+    id: 1,
+    cpf: "123.456.789-04",
+    nome: "Carlos Almeida",
+    matricula: "20250104",
+    status: "Indeferido",
+    progresso: 100,
+    documentos: 8,
+    dataInscricao: "2025-02-03T13:19:00",
+  },
+  {
+    id: 1,
+    cpf: "111.222.333-01",
+    nome: "Mariana Ferreira",
+    matricula: "20250107",
+    status: "Pendente",
+    progresso: 0,
+    documentos: 15,
+    dataInscricao: "2025-02-02T11:00:00",
+  },
+  {
+    id: 1,
+    cpf: "222.333.444-02",
+    nome: "Rafael Gonçalves",
+    matricula: "20250108",
+    status: "Deferido",
+    progresso: 100,
+    documentos: 30,
+    dataInscricao: "2025-02-01T10:15:00",
+  },
+];
