+++ src/routes/index.tsx
@@ -1,5 +1,4 @@
 import { createFileRoute } from '@tanstack/react-router'
-import { StudentDataTable } from '@/components/student-table'
 
 export const Route = createFileRoute('/')({
   component: App,
@@ -9,7 +8,7 @@ function App() {
   return (
     <div className="flex items-center justify-center h-screen">
       <div className=" w-full h-full flex items-center justify-center">
-        <StudentDataTable />
+        Home
       </div>
     </div>
   )
